feat(impact): show the full removal-impact report inside the delete confirmation (TEA-72)
Deleting an element no longer requires seeking out the Removal impact
dialog to see what breaks — the delete confirmation itself now renders
the full section-by-section report (goes away, breaks now, relationships
removed, orphans, views affected, downstream) whenever the deletion
targets are known. The report body is extracted into a shared
ImpactReportBody used by both dialogs; relationship, view, group, and
file deletions keep the compact confirm.
