+++ src/store/workspace-types.ts
@@ -22,6 +22,9 @@ export interface CascadeImpact {
 export interface PendingDelete {
   message: string
   impact?: CascadeImpact
+  /** Element ids being deleted — when set, the confirmation dialog renders the
+   *  full removal-impact report (what breaks, orphans, affected views) inline. */
+  targetIds?: string[]
   onConfirm: () => void
 }
 
@@ -59,7 +62,7 @@ export interface WorkspaceState extends UndoState {
   commandPaletteOpen: boolean
   pendingDelete: PendingDelete | null
   confirmDelete: (
-    payload: string | { message: string; impact?: CascadeImpact },
+    payload: string | { message: string; impact?: CascadeImpact; targetIds?: string[] },
     onConfirm: () => void,
   ) => void
   cancelDelete: () => void
